backend/requirements.txt
# Core
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.0
python-socketio==5.9.0
eventlet==0.33.3

# Serial communication
pyserial==3.5

# Data handling
pandas==2.0.0
numpy==1.24.0

# Optional: Advanced features
# python-engineio==4.5.1